refactor: 对齐 C11 项目模板：构建、脚本与命名重构
- 构建：引入 CMake presets，CMakeLists 拆分 CompilerOptions/Sources/Tests 模块
- 工具：新增 check/format/bootstrap 脚本（PowerShell + shell）
- 质量：新增 .clang-format/.clang-tidy/.gitattributes/.editorconfig/.clangd，
  全量重排并清零 clang-tidy 告警
- CI：改为真正运行 ctest
- 命名：类型统一 snake_case + _t（13 个）、函数 camelCase 改 snake_case（37 个）

+++ src/platform/platform_win.c
@@ -4,68 +4,59 @@
  */
 
 #include "platform.h"
-#include <windows.h>
 #include <io.h>
 #include <stdio.h>
+#include <windows.h>
 
 /* ========================================================================
  * 公共函数实现
  * ======================================================================== */
 
-int platform_init(void)
-{
+int platform_init(void) {
     /* Windows 平台初始化 - 当前无需特殊处理 */
     return 0;
 }
 
-void platform_cleanup(void)
-{
+void platform_cleanup(void) {
     /* Windows 平台清理 - 当前无需特殊处理 */
 }
 
-const char* platform_get_name(void)
-{
+const char *platform_get_name(void) {
     return "Windows";
 }
 
-uint64_t platform_get_tick_ms(void)
-{
+uint64_t platform_get_tick_ms(void) {
     /* 使用 Windows Performance Counter 获取高精度时间 */
     static LARGE_INTEGER frequency = {0};
     LARGE_INTEGER counter;
-    
+
     if (frequency.QuadPart == 0) {
         QueryPerformanceFrequency(&frequency);
     }
-    
+
     QueryPerformanceCounter(&counter);
-    
+
     /* 转换为毫秒 */
-    return (counter.QuadPart * 1000) / frequency.QuadPart;
+    return (uint64_t)((counter.QuadPart * 1000) / frequency.QuadPart);
 }
 
-void platform_sleep_ms(uint32_t ms)
-{
+void platform_sleep_ms(uint32_t ms) {
     Sleep(ms);
 }
 
-void platform_enable_utf8(void)
-{
+void platform_enable_utf8(void) {
     // 设置控制台代码页为 UTF-8 (65001)
     // 先判断是否在控制台中运行
-    if (_isatty(_fileno(stdout)))
-    {
+    if (_isatty(_fileno(stdout))) {
         UINT output_cp = GetConsoleOutputCP();
-        UINT input_cp  = GetConsoleCP();
+        UINT input_cp = GetConsoleCP();
 
         // 只有当不是 UTF-8 时才修改
-        if (output_cp != 65001)
-        {
+        if (output_cp != 65001) {
             SetConsoleOutputCP(65001);
         }
 
-        if (input_cp != 65001)
-        {
+        if (input_cp != 65001) {
             SetConsoleCP(65001);
         }
     }
